+++ backend/django_project/settings.py
@@ -72,6 +72,7 @@ INSTALLED_APPS = [
     # Local apps
     "users.apps.UsersConfig",
     "bookmarks.apps.BookmarksConfig",
+    "downloads.apps.DownloadsConfig",
 ]
 
 MIDDLEWARE = [
@@ -162,3 +163,15 @@ MEDIA_ROOT = BASE_DIR / "media"
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+# Redis
+REDIS_HOST = env.str("REDIS_HOST", "redis")
+REDIS_PORT = env.int("REDIS_PORT", 6379)
+REDIS_DB = env.int("REDIS_DB", 0)
+
+# Celery
+CELERY_BROKER_URL = "redis://redis:6379"
+
+# Telegram
+# These env vars must be set for worker service in docker-compose.yml
+TELEGRAM_BOT_TOKEN = env.str("TELEGRAM_BOT_TOKEN", None)